remove original Battery component
